Replaced inline checks
inline checks created from the party sheet have the party defined as the actor performing the roll, even when posted to chat, and no amount of clicking on player tokens will override this. To get around this, I replaced the inline check button in the subsystems tab with a button that posts a new inline link to chat, completely separated from the party sheet. Unfortunate that there's an intermediate button that the GM has to click (if you were planning on clicking it for your players), but it's the same number of clicks if players are the ones rolling.